+++ src/lib.rs
@@ -247,11 +247,12 @@ pub mod prelude {
     };
     #[cfg(feature = "sys")]
     pub use crate::sys::{
-        Capability, CapabilityView, HaNode, HaStatus, Health, LeaderStatus, LockedUsers,
-        LockedUsersMountAccessor, LockedUsersNamespace, LoggerLevel, LoggerLevels, NamespaceInfo,
-        NamespaceList, NamespaceRequest, RaftAutopilotConfig, RaftConfiguration, RaftJoinRequest,
-        RaftJoinResponse, RaftPeerRequest, RaftServer, RateLimitQuotaConfig, RateLimitQuotaInfo,
-        RateLimitQuotaList, RateLimitQuotaRequest, RemountMigrationInfo, RemountRequest,
-        RemountResponse, RemountStatus, Sys, VersionHistory, VersionHistoryEntry,
+        Capability, CapabilityView, CorsConfig, CorsConfigRequest, HaNode, HaStatus, Health,
+        KeyStatus, LeaderStatus, LockedUsers, LockedUsersMountAccessor, LockedUsersNamespace,
+        LoggerLevel, LoggerLevels, NamespaceInfo, NamespaceList, NamespaceRequest,
+        RaftAutopilotConfig, RaftConfiguration, RaftJoinRequest, RaftJoinResponse, RaftPeerRequest,
+        RaftServer, RateLimitQuotaConfig, RateLimitQuotaInfo, RateLimitQuotaList,
+        RateLimitQuotaRequest, RemountMigrationInfo, RemountRequest, RemountResponse,
+        RemountStatus, Sys, VersionHistory, VersionHistoryEntry,
     };
 }